% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{average_groups}
\alias{average_groups}
\title{Group Averages}
\usage{
average_groups(
  res,
  data = NULL,
  group,
  by = NULL,
  over = NULL,
  est = NULL,
  svy = NULL,
  user_na = FALSE,
  ...
)
}
\arguments{
\item{res}{(dataframe) Data frame of results, including estimates and standard errors to be averaged.}

\item{data}{(dataframe) Data frame containing replicated weights.}

\item{group}{(grp function) Function used to compute grouped averages; takes arguments group.name, column, and cases, and appends results at the end of the data frame.}

\item{by}{(string vector) Variable(s) used to break down the results.}

\item{over}{(vector string) Variable(s) over which the analysis is performed.}

\item{est}{(est function) Estimation function; takes arguments what = estimate, tgt = target, and rgr = regressor.}

\item{svy}{(string) Name of the survey/project to analyse (e.g. TALISSCH, TALISTCH).}

\item{user_na}{(bool) If TRUE, shows the nature of user-defined missing values for the grouping variable(s).}

\item{...}{Additional arguments such as na_to_zero: (Bool) TRUE → treats NA values as zero for simple average calculations.}
}
\value{
Dataframe with averages or weighted averages (totals) appended in the last rows for the selected cases.
}
\description{
Group Averages
}
