% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd_data.R
\docType{data}
\name{df_pisa18}
\alias{df_pisa18}
\title{PISA 2018 Student questionnaire Database}
\format{
A data frame with 1269 rows and 1120 variables
}
\usage{
data(df_pisa18)
}
\description{
This dataset is a subset of the PISA 2018 database produced by the OECD for the countries of France, Italy, and Mexico.
}
\keyword{datasets}
