% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd_data.R
\docType{data}
\name{df_talis18}
\alias{df_talis18}
\title{TALIS 2018 Teachers Database}
\format{
A data frame with 548 rows and 496 variables
}
\usage{
data(df_talis18)
}
\description{
This dataset is a subset of the TALIS 2018 database produced by the OECD for the countries of France, Italy, and Mexico.
}
\keyword{datasets}
