% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.data.categ.vars.R
\name{format_data_categ_vars}
\alias{format_data_categ_vars}
\title{Format categorical variables as factor for Rrepest}
\usage{
format_data_categ_vars(df, categ.vars, show_na = F)
}
\arguments{
\item{df}{(data frame) Data to analyse.}

\item{categ.vars}{(string vector) Categorical variables for analysis.}

\item{show_na}{(bool) Keeps NAs as categories to get frequency from.}
}
\value{
Data frame with categorical variables as factors for frequencies or over variables.
}
\description{
Format categorical variables as factor for Rrepest
}
\examples{
format_data_categ_vars(df = mtcars, categ.vars = "cyl") 
}
