% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_rrepest_lin_reg.R
\name{format_data_cont_vars}
\alias{format_data_cont_vars}
\title{Format continuous variables as numeric for Rrepest}
\usage{
format_data_cont_vars(df, cont.vars)
}
\arguments{
\item{df}{(data frame) Data to be analysed.}

\item{cont.vars}{(string vector) continuous variables for analysis}
}
\value{
Data frame with cantinuous variables converted to numeric for a continuous analysis (means, regression, etc.)
}
\description{
Format continuous variables as numeric for Rrepest
}
\examples{
format_data_cont_vars(mtcars,"hp")
}
