% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_rrepest_groups_argument.R
\name{grp}
\alias{grp}
\title{Group list}
\usage{
grp(group.name, column, cases, full_weight = FALSE)
}
\arguments{
\item{group.name}{(string) Name of the group to be displayed}

\item{column}{(string) Column/variable of interest to be grouped}

\item{cases}{(string vector) Rows/values to be included in the group}

\item{full_weight}{(bool) TRUE: average of the group will be weighted average}
}
\value{
List containing the arguments for the grp() function used within Rrepest() function's average() and group() options.
}
\description{
Obtains a list with the arguments for the grp() function used within Rrepest() function's average() and group() options.
}
\examples{
append(grp("OECD Average","CNTRY",c("HUN","MEX")), grp("Europe","CNTRY",c("ITA","FRA")))

}
