% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv_test.R
\name{inv_test}
\alias{inv_test}
\title{inv_test}
\usage{
inv_test(data, name_index)
}
\arguments{
\item{data}{(dataframe) df to analyze}

\item{name_index}{(string/numeric) name or index for the estimate (b.) columns containing the data for the test in Rrepest)}
}
\value{
Dataframe cointaining inverted test column names for "b." and "se." according to Rrepest structure and column multiplied by (-1) for "b."
}
\description{
Invert test column from Rrepest test = TRUE by name on "b." and "se." in the column name and by sign (*-1) on "b."
}
\examples{
inv_test(rrepest_pisa_age_gender, name_index = 6) 
}
