% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrepest_pisa_age_gender}
\alias{rrepest_pisa_age_gender}
\title{Rrepest table of results for PISA 2018 showing age and gender}
\format{
\subsection{\code{rrepest_pisa_age_gender}}{

A data frame in tibble format with 3 rows and 7 columns:
\describe{
\item{cnt}{Country ISO 3166-1 alpha-3 codes.}
\item{b.mean.age..Female}{Mean age for female students.}
\item{se.mean.age..Female}{Standar error of the mean age for female students.}
\item{b.mean.age..Male}{Mean age for male students.}
\item{se.mean.age..Male}{Standar error of the mean age for male students.}
\item{b.mean.age..(Female-Male)}{Difference of the mean age from female to male students.}
\item{se.mean.age..(Female-Male)}{Standard error of the difference of the mean age from female to male students.}
}
}
}
\source{
\url{https://www.oecd.org/en/data/datasets/pisa-2018-database.html#data}
}
\usage{
rrepest_pisa_age_gender
}
\description{
A table of results from Rrepest of the mean age of students broken down by gender for PISA 2018.
}
\keyword{datasets}
