% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rrepest_pisa_age_isced}
\alias{rrepest_pisa_age_isced}
\title{Rrepest table of results for PISA 2018 showing the age and completed schooling level of students' mothers}
\format{
\subsection{\code{rrepest_pisa_age_isced}}{

A data frame in tibble format with 3 rows and 7 columns:
\describe{
\item{cnt}{Country ISO 3166-1 alpha-3 codes.}
\item{b.mean.age..ISCED level 1}{Mean age of students whose mothers completed ISCED 1.}
\item{se.mean.age..ISCED level 1}{Standard error of the mean age of students whose mothers completed ISCED 1.}
\item{b.mean.age..ISCED level 2}{Mean age of students whose mothers completed ISCED 2.}
\item{se.mean.age..ISCED level 2}{Standard error of the mean age of students whose mothers completed ISCED 2.}
\item{b.mean.age..ISCED level 3A}{Mean age of students whose mothers completed ISCED 3A.}
\item{se.mean.age..ISCED level 3A}{Standard error of the mean age of students whose mothers completed ISCED 3A.}
\item{b.mean.age..ISCED level 3B, 3C}{Mean age of students whose mothers completed ISCED 3B/3C.}
\item{se.mean.age..ISCED level 3B, 3C}{Standard error of the mean age of students whose mothers completed ISCED 3B/3C.}
\item{b.mean.age..She did not complete  ISCED level 1}{Mean age of students whose mothers did not complete ISCED 1.}
\item{se.mean.age..She did not complete  ISCED level 1}{Standard error of the mean age of students whose mothers did not complete ISCED 1.}
\item{b.mean.age..(ISCED level 1-She did not complete  ISCED level 1)}{Mean age difference between ISCED 1 completed vs. non-ISCED 1 completed mothers.}
\item{se.mean.age..(ISCED level 1-She did not complete  ISCED level 1)}{Standard error of the mean age difference between ISCED 1 completed vs. non-ISCED 1 completed mothers.}
...
}
}
}
\source{
\url{https://www.oecd.org/en/data/datasets/pisa-2018-database.html#data}
}
\usage{
rrepest_pisa_age_isced
}
\description{
A table of results from Rrepest of the mean age of students broken down by
completed schooling level of students' mothers for PISA 2018.
}
\keyword{datasets}
