% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_quantile.R
\name{weighted.quant}
\alias{weighted.quant}
\title{Weighted Quantile}
\usage{
weighted.quant(x, w = rep(1, length(x)), q = 0.5)
}
\arguments{
\item{x}{(numeric vector) variable from where to get quantiles}

\item{w}{(numeric vector) vector of weights}

\item{q}{(numeric vector) From 0 to 1 (exclusive) for the quantile desired}
}
\value{
Weighted quantile of a numeric vector
}
\description{
Computation of weighted quantiles
}
\examples{
weighted.quant(x = mtcars$mpg, w = mtcars$wt,  q = seq(.1,.9,.1))

}
