% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sdev.R
\name{weighted.std}
\alias{weighted.std}
\title{Weighted Standard Deviation}
\usage{
weighted.std(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{(numeric vector) variable to analyze}

\item{w}{(numeric vector) vector of weights}

\item{na.rm}{(bool) if TRUE remove missing values.}
}
\value{
Scalar with Variance or Standard Deviation
}
\description{
Calculate the standard deviation of a numeric vector
}
\examples{
data(df_talis18)

weighted.std(df_talis18$TT3G02, df_talis18$TRWGT1)
}
