% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_variance.R
\name{weighted.var}
\alias{weighted.var}
\title{Weighted variance}
\usage{
weighted.var(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{(numeric vector) Variable of interest for which to compute the variance}

\item{w}{(numeric vector) Vector with weights}

\item{na.rm}{(bool) if TRUE: Excludes missing values before computing the variance}
}
\value{
Scalar containing the variance
}
\description{
Computes the weighted variance of a numeric vector.
}
\examples{
data(df_talis18) 

weighted.var(df_talis18$TT3G02, df_talis18$TRWGT1)
}
