% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/oecd_data.R
\docType{data}
\name{df_talis18}
\alias{df_talis18}
\title{Teaching and Learning International Survey (TALIS) 2018 noisy data subset}
\format{
\subsection{\code{df_talis18}}{

A data frame in tibble format with 496 rows and 548 columns from which only some relevant variables are listed below.
For a detailed description of all variables go to https://www.oecd.org/en/data/datasets/talis-2018-database.html#data
\describe{
\item{idlang}{Country language}
\item{cntry}{Country ISO 3166-1 alpha-3 codes.}
\item{tt3g01}{Gender}
\item{tt3g02}{Age}
\item{tt3g03}{Highest level of formal education completed}
...
}
}

A data frame with 548 rows/observations and 496 columns/variables.
}
\source{
\url{https://www.oecd.org/en/data/datasets/talis-2018-database.html#data}
}
\usage{
df_talis18

data(df_talis18)
}
\description{
A subset of noisy data from the Teaching and Learning International Survey (TALIS) 2018 cycle
for Mexico, Italy, and France.

This dataset is a subset of the TALIS 2018 database produced by the OECD for the countries of France, Italy, and Mexico.
}
\keyword{datasets}
