% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_rrepest_estimate_argument.R
\name{est}
\alias{est}
\title{Estimate list}
\usage{
est(statistic, target, regressor = NULL)
}
\arguments{
\item{statistic}{(string vector) accepts "mean","var","std", "quant", "iqr", "freq", "lm", "corr", "cov"}

\item{target}{(string vector) variable from where to get estimation}

\item{regressor}{(string vector) independent variable for regression (1+)}
}
\value{
list of components to estimate for repest
}
\description{
Input the statistic wanted, target variable, and (optional) list of regressors
}
\examples{
est(c("mean","quant",.5,"corr"),c("pv1math","pv1read","Pv1SCIE"))

}
