% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_rrepest_groups_argument.R
\name{grp}
\alias{grp}
\title{grp}
\usage{
grp(group.name, column, cases)
}
\arguments{
\item{group.name}{(string) Name of the group to be displayed}

\item{column}{(string) Column where the data is located}

\item{cases}{(string vector) List of values to be put into the group}
}
\value{
list of groups to redefine group_name = column, values_in_group
}
\description{
Obtain a list as argument for groups to be evaluated in data
}
\examples{
append(grp("OECD Average","CNTRY",c("HUN","MEX")), grp("Europe","CNTRY",c("ITA","FRA")))
}
