% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_iqr.R
\name{weighted.iqr}
\alias{weighted.iqr}
\title{Weighted Interquantile Range}
\usage{
weighted.iqr(x, w = rep(1, length(x)), rang = c(0.25, 0.75))
}
\arguments{
\item{x}{(numeric vector) variable from where to get quantiles}

\item{w}{(numeric vector) vector of weights}

\item{rang}{(numeric vector) two numbers indicating the range of the quantiles}
}
\value{
Interquantile range
}
\description{
Compute interquantile range
}
\examples{
weighted.iqr(x = mtcars$mpg, w = mtcars$wt,  rang = c(.5,.9))
  
}
