% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_variance.R
\name{weighted.var}
\alias{weighted.var}
\title{Weighted variance}
\usage{
weighted.var(x, w, na.rm = TRUE)
}
\arguments{
\item{x}{(numeric vector) variable to analyze}

\item{w}{(numeric vector) vector of weights}

\item{na.rm}{(bool) if TRUE remove missing values.}
}
\value{
Scalar with Variance or Standard Deviation
}
\description{
Calculate the weighted variance numeric vector
}
\examples{
data(df_talis18) 

weighted.var(df_talis18$TT3G02, df_talis18$TRWGT1)
}
